requirements.txt
# Azure SDK dependencies
azure-identity==1.15.0
azure-mgmt-resource==23.0.1
azure-mgmt-compute==30.5.0
azure-mgmt-network==25.2.0
azure-mgmt-storage==21.1.0
azure-mgmt-monitor==6.0.2
azure-mgmt-security==6.0.0

# AI/OpenAI dependencies
openai>=1.50.0,<2.0.0

# PowerPoint generation
python-pptx==0.6.23

# PDF generation
fpdf2==2.8.2

# Configuration and utilities
python-dotenv==1.0.1
pyyaml==6.0.1

# Data processing
pandas==2.2.0
